requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.2.4
django-crispy-forms==2.3
django-crontab==0.7.1
django-dbbackup==4.3.0
django-storages==1.14.6
webdav4==0.9.8
dotenv==0.9.9
psycopg2-binary==2.9.10
pypdftk==0.5
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
substring==0.2
whitenoise==5.0.1
uwsgi==2.0.26
psutil==7.0.0

# Testing dependencies
playwright==1.40.0
pytest-playwright==0.4.3
pytest-django==4.11.1
cpf-generator==2.1.4
